+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   - Adding sdx-common functionality
   - Updating logger format using sdx-common
   - Adding reprocessing functionality
+  - Updating reprocessing functionality, removing queuepublisher and encrypter
 
 ### 1.0.0 2016-08-16
   - Initial release